refactor: Move auth0 functions to a dedicated module (bug 1879094)
