+++ api/tests/test_settings.py
@@ -1,9 +1,5 @@
 import os
 
-import pytest
-
-from shipit_api import admin
-
 
 def flatten(lst):
     return [item for sublist in lst for item in sublist]
@@ -94,69 +90,3 @@ def test_scopes(monkeypatch):
     )
 
     assert scopes == expected_scopes
-
-
-@pytest.mark.parametrize(
-    "product_name, product_config, expected",
-    (
-        (
-            "can-be-disabled-product",
-            {"can-be-disabled": True, "repo-url": ""},
-            [
-                "add_release/can-be-disabled-product",
-                "abandon_release/can-be-disabled-product",
-                "disable_product/can-be-disabled-product",
-                "enable_product/can-be-disabled-product",
-            ],
-        ),
-        (
-            "product-on-github",
-            {"can-be-disabled": False, "repo-url": "https://github.com/some-org/product-on-github"},
-            [
-                "add_release/product-on-github",
-                "abandon_release/product-on-github",
-                "github",
-            ],
-        ),
-        (
-            "firefox",  # Must include rc phases
-            {
-                "can-be-disabled": True,
-                "authorized-ldap-groups": [
-                    "shipit_firefox",
-                    "shipit_relman",
-                ],
-                "phases": [
-                    "promote_firefox",
-                    "push_firefox",
-                    "ship_firefox",
-                ],
-                "firefox-ci-trust-domain": "gecko",
-                "repo-url": "",
-                "version-class": "mozilla_version.gecko.GeckoVersion",
-            },
-            [
-                "add_release/firefox",
-                "abandon_release/firefox",
-                "schedule_phase/firefox/promote_firefox",
-                "phase_signoff/firefox/promote_firefox",
-                "schedule_phase/firefox/push_firefox",
-                "phase_signoff/firefox/push_firefox",
-                "schedule_phase/firefox/ship_firefox",
-                "phase_signoff/firefox/ship_firefox",
-                "schedule_phase/firefox/promote_firefox_rc",
-                "phase_signoff/firefox/promote_firefox_rc",
-                "schedule_phase/firefox/ship_firefox_rc",
-                "phase_signoff/firefox/ship_firefox_rc",
-                "schedule_phase/firefox/push_firefox",
-                "phase_signoff/firefox/push_firefox",
-                "schedule_phase/firefox/ship_firefox",
-                "phase_signoff/firefox/ship_firefox",
-                "disable_product/firefox",
-                "enable_product/firefox",
-            ],
-        ),
-    ),
-)
-def test_get_auth0_scopes(product_name, product_config, expected):
-    assert admin.settings._get_auth0_scopes(product_name, product_config) == expected
